Add partner parameter to auth URL
This patch adds a `partner=docker` to any auth URL obtained from `tailscale up`.
I know that using the `URL` class is kind of overkill, but previous experience
has taught me that it's better not to make assumptions about the existing structure of
the URL, and to leverage JS to just handle that for us.

Signed-off-by: Aaron Klotz <aaron@tailscale.com>