Fix user ID truncation bug
For a couple members of the Docker team, the extension wasn't showing
their user information. It turns out, `tailscale status --json` returns
the UserID as a number in some places, which results in truncation of
large IDs. This means our extension code can't find the right user
information in the users map.

This commit fixes it by parsing out a string value from the raw textual
output of `tailscale status --json` if we can't find the user in the
users map.