Add MagicDNS support
This commit adds MagicDNS support by looking for when MagicDNS is
enabled on the tailnet, and replacing IP addresses with DNS names in
the Tailscale IP column.

It also adds a new "more" option to the table which allows users to copy
the IP address, even if MagicDNS is enabled.