Lower polling frequency on onboarding/authorization views
We were previously polling every 750ms for Tailscale's status and
Docker's container state. We did this to stay responsive to when users
log in or when their devices get authorized.

However, this may be the result of some performance issues in some
circumstances (unsure, still gathering details). In any case, with the
addition of focus-based refetching in ebc4e163, the need for low-latency
polling is reduced a bit, so I'm lowering this down 2.5s.