Use Docker events to update container list
This commit introduces a global event emitter for Docker events, which
filters events to non-Tailscale extension events. When a `container` or
`network` event comes in, we re-fetch the list of containers. And since
events sometimes get triggered in quick succession, we debounce the
fetching so we only get a single update at a time.

This should greatly reduce the performance impact of container fetching.