Update CSS to defensively avoid inheriting Docker styles
This commit opts us out of Docker's default theme which is designed for
MUI, and seems like it may change a fair bit. Instead, we update our
Tailwind config to inherit a selection of CSS variables from their theme
so we have greater control over changes and can avoid bitrot.