Fix local extension development
Due to a Docker Desktop issue (docker/extensions-sdk#215), all local
extension containers need to have a namespace prefix before a slash (eg.
tailscale/my-extension). Failure to do so can cause issues when trying
to remove extensions.

While the bug might be fixed, it seems like this prefix is recommended
anyways, so we'll keep it around for the long term.